sessions: split session model from view service
Separate the session model (ISessionsManagementService) from the view
(ISessionsViewService) so the model owns canonical session state and the
view owns visible-slot arrangement, focus and per-session view
persistence. Introduces the new `browser/sessionsViewService.ts` core
service and `browser/parts/sessionsParts.ts`, makes the part a passive
renderer, and adds support for opening a session without focusing it
(reactive preserve-focus intent published atomically with the active
session).

Also encapsulates the in-progress new-session draft in the model as a
reactive `newSession` observable with a conditional `discardNewSession`,
merges `openNewSessionView` into `openNewSession`, and routes new-session
creation through the view so the draft is activated (fixing the session
type picker rendering empty after reload).
