+++ .github/skills/sessions/SKILL.md
@@ -33,9 +33,7 @@ Then read the relevant spec for the area you are changing (see table below). If 
 - **Modifying workbench code**: Prefer extending/wrapping workbench classes in the sessions layer over modifying shared workbench components.
 - **Timeouts as fixes**: Never use `setTimeout`/`disposableTimeout`/arbitrary delays to fix bugs or implement behaviour. They are race-prone guesses that mask the real ordering/state problem. Drive logic off deterministic signals instead — observables (`autorun`/`derived`), explicit events (`onDidChange*`), lifecycle phases, or awaiting the actual async operation.
 - **Stashed state read back later (side-channels)**: Never stash a value on a service during one method call and read it back from a separate query later, assuming it is still valid (e.g. a `Set`/flag set in `openSession` and consumed by a `shouldX()` pull-API). This is fragile temporal coupling. Instead, make it reactive state that is set **atomically together with its source of truth** and consumed reactively. Example: per-activation intent like "open in background / preserve focus" is exposed as an `IObservable` set in the **same transaction** as `activeSession` (via a single internal setter so it can never go stale), and read with `.read(reader)` in the consumer's `autorun` — never via a consume-once getter.
-- **Blocking on a "pending/waiting" state instead of creating + upgrading**: When an entity (e.g. a draft session) depends on something that registers asynchronously, don't withhold creation behind a pending/waiting state. Prefer creating immediately with the best available data, then **replace/upgrade** it once the awaited dependency arrives (driven by an `onDidChange*`/observable signal), cancelling the upgrade if the user changes the inputs meanwhile. Do **not** bound the upgrade with a timeout or even a lifecycle milestone like `LifecyclePhase.Eventually` — an agent host connects lazily and can surface its session types after `Eventually` has already fired, which would lock in the wrong fallback. Let the upgrade listener live for the consumer's lifetime instead.
-- **Persisting a picker value only on manual selection**: A picker that writes storage only when the user actively changes the dropdown will lose any auto-selected/defaulted value on reload (the stored preference is empty, so it falls back to a default). Persist on **every** change of the underlying value — exactly like a normal picker — so the last effective value always survives reload. Example: `SessionTypePicker` writes `{ providerId, sessionTypeId }` from both the manual pick handler and the active-session `refresh`, through the single `_writeStoredPick` persistence point.
-- **Pre-validating a stored preference against not-yet-ready state in an event handler**: Don't re-check a restored preference against currently-available data (e.g. validating the stored session-type pick against `getSessionTypesForFolder` inside `onDidSelectWorkspace`) and null it out on mismatch. During reload the awaited provider hasn't registered yet, so a perfectly valid preference is discarded and the wrong default is locked in. Pass the preference straight through to the single place that reconciles preference-vs-availability (which creates now and upgrades later).
+- **Blocking on a "pending/waiting" state instead of creating + upgrading**: When an entity (e.g. a draft session) depends on something that registers asynchronously, don't withhold creation behind a pending/waiting state. Prefer creating immediately with the best available data, then **replace/upgrade** it once the awaited dependency arrives (driven by an `onDidChange*`/observable signal), cancelling the upgrade if the user changes the inputs meanwhile and giving up at a deterministic lifecycle milestone (e.g. `LifecyclePhase.Eventually`) rather than a timeout. This keeps the UI populated and avoids fragile "is everything ready yet?" gating.
 
 ## Capturing Feedback (meta-rule)
