+++ src/vs/sessions/contrib/layout/browser/sessionLayoutController.ts
@@ -23,6 +23,7 @@ import { IWorkbenchLayoutService, Parts } from '../../../../workbench/services/l
 import { IPaneCompositePartService } from '../../../../workbench/services/panecomposite/browser/panecomposite.js';
 import { IViewsService } from '../../../../workbench/services/views/common/viewsService.js';
 import { IActiveSession, ISessionsManagementService } from '../../../services/sessions/common/sessionsManagement.js';
+import { ISessionsViewService } from '../../../browser/sessionsViewService.js';
 import { SessionStatus } from '../../../services/sessions/common/session.js';
 import { CHANGES_VIEW_ID } from '../../changes/common/changes.js';
 import { SESSIONS_FILES_CONTAINER_ID } from '../../files/browser/files.contribution.js';
@@ -76,6 +77,7 @@ export class LayoutController extends Disposable {
 	constructor(
 		@IWorkbenchLayoutService private readonly _layoutService: IWorkbenchLayoutService,
 		@ISessionsManagementService private readonly _sessionManagementService: ISessionsManagementService,
+		@ISessionsViewService private readonly _sessionsViewService: ISessionsViewService,
 		@IChatService private readonly _chatService: IChatService,
 		@IViewsService private readonly _viewsService: IViewsService,
 		@IPaneCompositePartService private readonly _paneCompositePartService: IPaneCompositePartService,
@@ -122,7 +124,7 @@ export class LayoutController extends Disposable {
 		});
 
 		const multipleSessionsVisibleObs = derived<boolean>(reader => {
-			return this._sessionManagementService.visibleSessions.read(reader).length > 1;
+			return this._sessionsViewService.visibleSessions.read(reader).length > 1;
 		});
 
 		// When multiple sessions are visible, drop per-session view/panel state
@@ -130,7 +132,7 @@ export class LayoutController extends Disposable {
 		// This will ensure the default visibility logic will be used again after
 		// closing all visible session and opening an existing one
 		this._register(autorun(reader => {
-			const visibleSessions = this._sessionManagementService.visibleSessions.read(reader);
+			const visibleSessions = this._sessionsViewService.visibleSessions.read(reader);
 			if (visibleSessions.length <= 1) {
 				return;
 			}
@@ -426,7 +428,7 @@ export class LayoutController extends Disposable {
 
 	private _saveState(): void {
 		const activeSession = this._sessionManagementService.activeSession.get();
-		const multipleVisible = this._sessionManagementService.visibleSessions.get().length > 1;
+		const multipleVisible = this._sessionsViewService.visibleSessions.get().length > 1;
 
 		// Capture current state for the active session (skip when multiple sessions are visible)
 		if (activeSession && !multipleVisible) {
