+++ src/vs/sessions/contrib/layout/test/browser/sessionLayoutController.test.ts
@@ -25,6 +25,7 @@ import { IPaneCompositePartService } from '../../../../../workbench/services/pan
 import { IPaneComposite } from '../../../../../workbench/common/panecomposite.js';
 import { IViewsService } from '../../../../../workbench/services/views/common/viewsService.js';
 import { IActiveSession, ISessionsChangeEvent, ISessionsManagementService } from '../../../../services/sessions/common/sessionsManagement.js';
+import { ISessionsViewService } from '../../../../browser/sessionsViewService.js';
 import { IChat, ISessionFileChange, ISessionWorkspace, SessionStatus } from '../../../../services/sessions/common/session.js';
 import { LayoutController } from '../../browser/sessionLayoutController.js';
 import { CHANGES_VIEW_ID } from '../../../changes/common/changes.js';
@@ -137,9 +138,11 @@ suite('LayoutController', () => {
 
 		instaService.stub(ISessionsManagementService, new class extends mock<ISessionsManagementService>() {
 			override activeSession = activeSessionObs;
-			override readonly visibleSessions = constObservable([]);
 			override readonly onDidChangeSessions = onDidChangeSessions.event;
 			override getSessions() { return []; }
+		});
+		instaService.stub(ISessionsViewService, new class extends mock<ISessionsViewService>() {
+			override readonly visibleSessions = constObservable([]);
 		});
 
 		onDidSubmitRequest = store.add(new Emitter<{ chatSessionResource: URI }>());
@@ -464,9 +467,11 @@ suite('LayoutController', () => {
 		const activeSession = observableValue<IActiveSession | undefined>('active', undefined);
 		instaService.stub(ISessionsManagementService, new class extends mock<ISessionsManagementService>() {
 			override activeSession = activeSession;
-			override readonly visibleSessions = constObservable([]);
 			override readonly onDidChangeSessions = Event.None;
 			override getSessions() { return []; }
+		});
+		instaService.stub(ISessionsViewService, new class extends mock<ISessionsViewService>() {
+			override readonly visibleSessions = constObservable([]);
 		});
 		instaService.stub(IChatService, new class extends mock<IChatService>() {
 			override readonly onDidSubmitRequest = Event.None;
