+++ src/vs/sessions/contrib/sessions/browser/sessionsActions.ts
@@ -20,6 +20,7 @@ import { SessionsCategories } from '../../../common/categories.js';
 import { CanGoBackContext, CanGoForwardContext, ChatSessionProviderIdContext, MultipleSessionsVisibleContext, SessionIsCreatedContext, SessionIsMaximizedContext, SessionIsStickyContext, SessionsFocusContext, SessionSupportsMultipleChatsContext, SessionsWelcomeVisibleContext } from '../../../common/contextkeys.js';
 import { ANY_AGENT_HOST_PROVIDER_RE } from '../../../common/agentHostSessionsProvider.js';
 import { IActiveSession, ISessionsManagementService } from '../../../services/sessions/common/sessionsManagement.js';
+import { ISessionsViewService } from '../../../browser/sessionsViewService.js';
 import { ISession } from '../../../services/sessions/common/session.js';
 import { ISessionsPartService } from '../../../browser/parts/sessionsPartService.js';
 import { ISessionsListModelService } from '../../../services/sessions/browser/sessionsListModelService.js';
@@ -46,11 +47,12 @@ registerAction2(class ShowSessionsPickerAction extends Action2 {
 
 	override async run(accessor: ServicesAccessor) {
 		const sessionsManagementService = accessor.get(ISessionsManagementService);
+		const sessionsViewService = accessor.get(ISessionsViewService);
 		const quickInputService = accessor.get(IQuickInputService);
 		const sessionsPartService = accessor.get(ISessionsPartService);
 		const sessionsListModelService = accessor.get(ISessionsListModelService);
 
-		const { recent, other } = sessionsManagementService.getRecentlyOpenedSessions();
+		const { recent, other } = sessionsViewService.getRecentlyOpenedSessions();
 		const recentSessions = recent.filter(s => !s.isArchived.get());
 		const otherSessions = other.filter(s => !s.isArchived.get());
 
@@ -146,7 +148,7 @@ registerAction2(class ShowSessionsPickerAction extends Action2 {
 
 		const openSelected = (selected: ISessionPickItem, inBackground: boolean, toSide: boolean): void => {
 			if (!selected.session) {
-				sessionsManagementService.openNewSessionView({ inheritWorkspaceFromActiveSession: true });
+				sessionsViewService.openNewSession({ fromActiveSession: true });
 				sessionsPartService.focusSession(sessionsManagementService.activeSession.get());
 				return;
 			}
@@ -156,9 +158,9 @@ registerAction2(class ShowSessionsPickerAction extends Action2 {
 			// normal open when there is no active session to anchor against or the
 			// session is already the active one.
 			if (toSide && activeSessionId !== undefined && selected.session.sessionId !== activeSessionId) {
-				sessionsManagementService.insertAt(selected.session, activeSessionId, 'right', !inBackground);
+				sessionsViewService.insertAt(selected.session, activeSessionId, 'right', !inBackground);
 			} else {
-				sessionsManagementService.openSession(selected.session.resource, { preserveFocus: inBackground });
+				sessionsViewService.openSession(selected.session.resource, { preserveFocus: inBackground });
 			}
 		};
 
@@ -217,8 +219,7 @@ registerAction2(class GoBackAction extends Action2 {
 	}
 
 	override async run(accessor: ServicesAccessor): Promise<void> {
-		const sessionsManagementService = accessor.get(ISessionsManagementService);
-		await sessionsManagementService.openPreviousSession();
+		await accessor.get(ISessionsViewService).openPreviousSession();
 	}
 });
 
@@ -259,8 +260,7 @@ registerAction2(class GoForwardAction extends Action2 {
 	}
 
 	override async run(accessor: ServicesAccessor): Promise<void> {
-		const sessionsManagementService = accessor.get(ISessionsManagementService);
-		await sessionsManagementService.openNextSession();
+		await accessor.get(ISessionsViewService).openNextSession();
 	}
 });
 
@@ -311,16 +311,16 @@ for (let index = 0; index < 9; index++) {
 		}
 
 		override async run(accessor: ServicesAccessor): Promise<void> {
-			const sessionsManagementService = accessor.get(ISessionsManagementService);
+			const sessionsViewService = accessor.get(ISessionsViewService);
 			const sessionsPartService = accessor.get(ISessionsPartService);
 
-			const visible = sessionsManagementService.visibleSessions.get();
+			const visible = sessionsViewService.visibleSessions.get();
 			if (index >= visible.length) {
 				return;
 			}
 
 			const session = visible[index];
-			sessionsManagementService.setActive(session);
+			sessionsViewService.setActive(session);
 			sessionsPartService.focusSession(session);
 		}
 	});
@@ -346,7 +346,7 @@ registerAction2(class CloseAllSessionsAction extends Action2 {
 	}
 
 	override async run(accessor: ServicesAccessor): Promise<void> {
-		accessor.get(ISessionsManagementService).closeAllSessions();
+		accessor.get(ISessionsViewService).closeAllSessions();
 	}
 });
 
@@ -370,8 +370,9 @@ registerAction2(class AddChatToSessionBarAction extends Action2 {
 			return;
 		}
 		const sessionsManagementService = accessor.get(ISessionsManagementService);
+		const sessionsViewService = accessor.get(ISessionsViewService);
 		const sessionsPartService = accessor.get(ISessionsPartService);
-		await sessionsManagementService.openNewChatInSession(session);
+		await sessionsViewService.openNewChatInSession(session);
 		sessionsPartService.focusSession(sessionsManagementService.activeSession.get());
 	}
 });
@@ -400,7 +401,7 @@ registerAction2(class TogglePinSessionAction extends Action2 {
 		if (!session) {
 			return;
 		}
-		accessor.get(ISessionsManagementService).toggleSessionStickiness(session);
+		accessor.get(ISessionsViewService).toggleSessionStickiness(session);
 	}
 });
 
@@ -462,9 +463,10 @@ registerAction2(class CloseSessionAction extends Action2 {
 
 	override async run(accessor: ServicesAccessor, session: IActiveSession | undefined): Promise<void> {
 		const sessionsManagementService = accessor.get(ISessionsManagementService);
+		const sessionsViewService = accessor.get(ISessionsViewService);
 		const sessionsPartService = accessor.get(ISessionsPartService);
 
-		sessionsManagementService.closeSession(session);
+		sessionsViewService.closeSession(session);
 		sessionsPartService.focusSession(sessionsManagementService.activeSession.get());
 	}
 });
@@ -491,6 +493,6 @@ registerAction2(class ToggleMaximizeSessionViewAction extends Action2 {
 
 	override async run(accessor: ServicesAccessor, session: IActiveSession | undefined): Promise<void> {
 		accessor.get(ISessionsPartService).toggleMaximizeSession(session);
-		accessor.get(ISessionsManagementService).setActive(session);
+		accessor.get(ISessionsViewService).setActive(session);
 	}
 });
