+++ src/vs/sessions/services/sessions/browser/sessionNavigation.ts
@@ -9,12 +9,22 @@ import { URI } from '../../../../base/common/uri.js';
 import { IContextKey, IContextKeyService } from '../../../../platform/contextkey/common/contextkey.js';
 import { ILogService } from '../../../../platform/log/common/log.js';
 import { CanGoBackContext, CanGoForwardContext } from '../../../common/contextkeys.js';
-import { SessionStatus } from '../common/session.js';
+import { ISession, SessionStatus } from '../common/session.js';
 import { ISessionsChangeEvent, ISessionsManagementService } from '../common/sessionsManagement.js';
 import { IRecencyEntry, SessionsRecencyHistory } from './sessionsRecencyHistory.js';
 
 function entryKey(sessionResource: URI, chatResource: URI | undefined): string {
 	return `${sessionResource.toString()}::${chatResource?.toString() ?? ''}`;
+}
+
+/**
+ * The subset of opening behaviour {@link SessionsNavigation} drives. Implemented
+ * by the view service, passed in to avoid the navigation (a `services` module)
+ * depending on the core view service.
+ */
+export interface ISessionOpener {
+	openSession(sessionResource: URI, options?: { preserveFocus?: boolean }): Promise<void>;
+	openChat(session: ISession, chatResource: URI): Promise<void>;
 }
 
 /**
@@ -67,6 +77,7 @@ export class SessionsNavigation extends Disposable {
 	});
 
 	constructor(
+		private readonly _opener: ISessionOpener,
 		private readonly _sessionsManagementService: ISessionsManagementService,
 		private readonly _recency: SessionsRecencyHistory,
 		contextKeyService: IContextKeyService,
@@ -195,12 +206,12 @@ export class SessionsNavigation extends Disposable {
 				if (entry.chatResource) {
 					const chatExists = session.chats.get().some(c => c.resource.toString() === entry.chatResource!.toString());
 					if (chatExists) {
-						await this._sessionsManagementService.openChat(session, entry.chatResource);
+						await this._opener.openChat(session, entry.chatResource);
 					} else {
-						await this._sessionsManagementService.openSession(entry.sessionResource);
+						await this._opener.openSession(entry.sessionResource);
 					}
 				} else {
-					await this._sessionsManagementService.openSession(entry.sessionResource);
+					await this._opener.openSession(entry.sessionResource);
 				}
 			} else {
 				// Session no longer exists, remove its entries from history
