+++ src/vs/sessions/services/sessions/browser/sessionsManagementService.ts
@@ -4,63 +4,21 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { Emitter, Event } from '../../../../base/common/event.js';
-import { CancellationToken, CancellationTokenSource } from '../../../../base/common/cancellation.js';
-import { disposableTimeout } from '../../../../base/common/async.js';
-import { Disposable, DisposableMap, DisposableStore, IDisposable, MutableDisposable } from '../../../../base/common/lifecycle.js';
-import { ResourceMap } from '../../../../base/common/map.js';
-import { IObservable, autorun } from '../../../../base/common/observable.js';
+import { Disposable, DisposableMap, DisposableStore, IDisposable } from '../../../../base/common/lifecycle.js';
+import { IObservable, autorun, observableValue } from '../../../../base/common/observable.js';
 import { URI } from '../../../../base/common/uri.js';
 import { IContextKey, IContextKeyService } from '../../../../platform/contextkey/common/contextkey.js';
 import { ILogService } from '../../../../platform/log/common/log.js';
-import { IStorageService, StorageScope, StorageTarget } from '../../../../platform/storage/common/storage.js';
-import { IAgentSessionsService } from '../../../../workbench/contrib/chat/browser/agentSessions/agentSessionsService.js';
 import { IChatService } from '../../../../workbench/contrib/chat/common/chatService/chatService.js';
 import { ChatAgentLocation } from '../../../../workbench/contrib/chat/common/constants.js';
 import { IChatWidgetHistoryService } from '../../../../workbench/contrib/chat/common/widget/chatWidgetHistoryService.js';
 import { IUriIdentityService } from '../../../../platform/uriIdentity/common/uriIdentity.js';
 import { ActiveSessionProviderIdContext, ActiveSessionTypeContext, IsActiveSessionArchivedContext, ActiveSessionWorkspaceIsVirtualContext, IsNewChatSessionContext } from '../../../common/contextkeys.js';
-import { ActiveSessionSupportsMultiChatContext, IActiveSession, ICreateNewSessionOptions, IProviderSessionType, IRecentlyOpenedSessions, ISendRequestOptions, ISendRequestSentEvent, ISessionsChangeEvent, ISessionsManagementService, IToggleSessionStickinessEvent } from '../common/sessionsManagement.js';
+import { ActiveSessionSupportsMultiChatContext, IActiveSession, ICreateNewSessionOptions, IProviderSessionType, ISendRequestOptions, ISendRequestSentEvent, ISessionsChangeEvent, ISessionsManagementService } from '../common/sessionsManagement.js';
 import { ISessionsProvidersChangeEvent, ISessionsProvidersService } from './sessionsProvidersService.js';
 import { ISessionChangeEvent, ISessionsProvider } from '../common/sessionsProvider.js';
 import { IChat, ISession, ISessionWorkspace, SessionStatus, ISessionType } from '../common/session.js';
 import { InstantiationType, registerSingleton } from '../../../../platform/instantiation/common/extensions.js';
-import { SessionsNavigation } from './sessionNavigation.js';
-import { SessionsRecencyHistory } from './sessionsRecencyHistory.js';
-import { VisibleSessions } from './visibleSessions.js';
-
-const ACTIVE_SESSION_STATES_KEY = 'agentSessions.activeSessionStates';
-
-/**
- * Upper bound on how long restore waits for a persisted session to resurface
- * via its provider. Generous (providers may load after auth settles) but finite
- * so a session that is gone for good cannot keep restore — and its provider
- * listeners — alive indefinitely.
- */
-const RESTORE_SESSION_WAIT_TIMEOUT = 30_000;
-
-/** Maximum number of recently opened sessions reported by {@link SessionsManagementService.getRecentlyOpenedSessions}. */
-const MAX_RECENTLY_OPENED_SESSIONS = 10;
-
-/**
- * Persisted state for a session.
- * Extend this interface to store additional per-session state that should be
- * remembered across restarts.
- */
-interface ISessionState {
-	/** The resource URI of the session. */
-	sessionResource: string;
-	/** The resource URI of the last active chat within the session. */
-	activeChatResource?: string;
-	/** Whether this session was the active session at the time of save. */
-	isActive?: boolean;
-	/**
-	 * Position (left-to-right) of the session in the grid at save time, when
-	 * the session was visible. `undefined` when the session was not visible.
-	 */
-	visibleOrder?: number;
-	/** Whether the session was pinned (sticky) in the grid at save time. */
-	isSticky?: boolean;
-}
 
 export class SessionsManagementService extends Disposable implements ISessionsManagementService {
 
@@ -87,48 +45,29 @@ export class SessionsManagementService extends Disposable implements ISessionsMa
 	private readonly _onDidRenameChat = this._register(new Emitter<ISession>());
 	readonly onDidRenameChat: Event<ISession> = this._onDidRenameChat.event;
 
-	private readonly _onDidToggleSessionStickiness = this._register(new Emitter<IToggleSessionStickinessEvent>());
-	readonly onDidToggleSessionStickiness: Event<IToggleSessionStickinessEvent> = this._onDidToggleSessionStickiness.event;
-
 	private readonly _onDidChangeSessionTypes = this._register(new Emitter<void>());
 	readonly onDidChangeSessionTypes: Event<void> = this._onDidChangeSessionTypes.event;
 
+	private readonly _onDidReplaceSession = this._register(new Emitter<{ readonly from: ISession; readonly to: ISession }>());
+	readonly onDidReplaceSession: Event<{ readonly from: ISession; readonly to: ISession }> = this._onDidReplaceSession.event;
+
 	private _sessionTypes: readonly ISessionType[] = [];
 
-	/** Owns the active/sticky/transient visibility model and the {@link ActiveSession} wrappers. */
-	private readonly _visibility: VisibleSessions;
-	readonly activeSession: IObservable<IActiveSession | undefined>;
-	readonly visibleSessions: IObservable<readonly (IActiveSession | undefined)[]>;
+	/** The canonical active session, mirrored from the visible active slot by the view service. */
+	private readonly _activeSession = observableValue<IActiveSession | undefined>(this, undefined);
+	readonly activeSession: IObservable<IActiveSession | undefined> = this._activeSession;
 
-	/** Tracks the pending new session so it can be restored by {@link openNewSessionView}. */
-	private _pendingNewSession: ISession | undefined;
+	/** Tracks the in-progress new session (composed but not yet sent). */
+	private readonly _newSession = observableValue<ISession | undefined>(this, undefined);
+	readonly newSession: IObservable<ISession | undefined> = this._newSession;
+
 	private readonly _isNewChatSessionContext: IContextKey<boolean>;
 	private readonly _activeSessionProviderId: IContextKey<string>;
 	private readonly _activeSessionType: IContextKey<string>;
 	private readonly _activeSessionWorkspaceIsVirtual: IContextKey<boolean>;
 	private readonly _isActiveSessionArchived: IContextKey<boolean>;
 	private readonly _supportsMultiChat: IContextKey<boolean>;
-	/** Cancelled on every navigation action so in-flight async opens bail out. */
-	private readonly _openSessionCts = this._register(new MutableDisposable<CancellationTokenSource>());
-	/**
-	 * Cancellation for the in-flight {@link restoreVisibleSessions}. Kept
-	 * separate from {@link _openSessionCts} so that additive new-session
-	 * operations (the new-chat composer eagerly creating a draft on startup)
-	 * do not abort restoring the previously visible grid. Only an explicit
-	 * navigation to a specific session cancels a restore.
-	 */
-	private readonly _restoreCts = this._register(new MutableDisposable<CancellationTokenSource>());
-	private readonly _onDidOpenNewSessionView = this._register(new Emitter<void>());
 	private readonly _providerListeners = this._register(new DisposableMap<string, IDisposable>());
-	private readonly _sessionStates: ResourceMap<ISessionState>;
-	private readonly _navigation: SessionsNavigation;
-	/**
-	 * The single source of truth for session recency (most-recently-opened
-	 * first), persisted across restarts. Both the recent-sessions picker (via
-	 * {@link getRecentlyOpenedSessions}) and {@link SessionsNavigation} build on
-	 * top of it.
-	 */
-	private readonly _recencyHistory: SessionsRecencyHistory;
 
 	/**
 	 * Chat resources for which this service has just kicked off a
@@ -140,12 +79,10 @@ export class SessionsManagementService extends Disposable implements ISessionsMa
 	private readonly _pendingSendChatResources = new Set<string>();
 
 	constructor(
-		@IStorageService private readonly storageService: IStorageService,
 		@ILogService private readonly logService: ILogService,
 		@IContextKeyService contextKeyService: IContextKeyService,
 		@ISessionsProvidersService private readonly sessionsProvidersService: ISessionsProvidersService,
 		@IUriIdentityService private readonly uriIdentityService: IUriIdentityService,
-		@IAgentSessionsService private readonly agentSessionsService: IAgentSessionsService,
 		@IChatService private readonly chatService: IChatService,
 		@IChatWidgetHistoryService private readonly chatWidgetHistoryService: IChatWidgetHistoryService,
 	) {
@@ -160,22 +97,6 @@ export class SessionsManagementService extends Disposable implements ISessionsMa
 		this._isActiveSessionArchived = IsActiveSessionArchivedContext.bindTo(contextKeyService);
 		this._supportsMultiChat = ActiveSessionSupportsMultiChatContext.bindTo(contextKeyService);
 
-		// Load persisted state
-		this._sessionStates = this._loadSessionStates();
-
-		// Visibility model — owns wrappers, active/sticky/transient state, and
-		// observables exposed to the UI.
-		this._visibility = this._register(new VisibleSessions(
-			session => this._restoreInitialChat(session),
-			this.uriIdentityService,
-			this.agentSessionsService,
-		));
-		this.activeSession = this._visibility.activeSession;
-		this.visibleSessions = this._visibility.visibleSessions;
-
-		// Save on shutdown
-		this._register(this.storageService.onWillSaveState(() => this._saveSessionStates()));
-
 		// Subscribe to provider changes for session type updates
 		this._register(this.sessionsProvidersService.onDidChangeProviders(e => {
 			this._onProvidersChanged(e);
@@ -184,26 +105,10 @@ export class SessionsManagementService extends Disposable implements ISessionsMa
 		this._subscribeToProviders(this.sessionsProvidersService.getProviders());
 		this._sessionTypes = this._collectSessionTypes();
 
-		// Session recency history — the single source of truth for "recently
-		// opened" ordering, shared by the picker and navigation.
-		this._recencyHistory = this._register(new SessionsRecencyHistory(
-			this.storageService,
-			this.logService,
-		));
-
-		// Session navigation history (Back/Forward) builds on the recency history.
-		this._navigation = this._register(new SessionsNavigation(
-			this,
-			this._recencyHistory,
-			contextKeyService,
-			this.logService,
-		));
-		this._register(this.onDidChangeSessions(e => this._navigation.onDidRemoveSessions(e)));
-		this._register(this.onDidDeleteSession(session => this._recencyHistory.remove(entry => entry.sessionResource.toString() === session.resource.toString())));
-
 		this._register(autorun(reader => {
-			const activeSession = this._visibility.activeSession.read(reader);
-			this._handleActiveSessionContextKeys(activeSession);
+			const activeSession = this._activeSession.read(reader);
+			const newSession = this._newSession.read(reader);
+			this._handleActiveSessionContextKeys(activeSession, newSession);
 			if (activeSession) {
 				reader.store.add(this._activeSessionListeners(activeSession));
 			}
@@ -236,13 +141,13 @@ export class SessionsManagementService extends Disposable implements ISessionsMa
 		}));
 	}
 
-	private _handleActiveSessionContextKeys(session: IActiveSession | undefined): void {
+	private _handleActiveSessionContextKeys(session: IActiveSession | undefined, newSession: ISession | undefined): void {
 		// Update context keys from session data
 		// IsNewChatSessionContext is true when no active session exists, OR when the
-		// active session is still pending (created but not yet sent for the first time).
-		// Scoping to the active session avoids flipping into "new chat" mode while
-		// viewing a different established session.
-		this._isNewChatSessionContext.set(session === undefined || session.sessionId === this._pendingNewSession?.sessionId);
+		// active session is still the in-progress new session (created but not yet
+		// sent for the first time). Scoping to the active session avoids flipping
+		// into "new chat" mode while viewing a different established session.
+		this._isNewChatSessionContext.set(session === undefined || session.sessionId === newSession?.sessionId);
 		this._activeSessionProviderId.set(session?.providerId ?? '');
 		this._activeSessionType.set(session?.sessionType ?? '');
 		this._activeSessionWorkspaceIsVirtual.set(session?.workspace.get()?.isVirtualWorkspace ?? true);
@@ -254,50 +159,14 @@ export class SessionsManagementService extends Disposable implements ISessionsMa
 		const disposables = new DisposableStore();
 
 		// Track archived state changes for the active session
-		let wasArchived = activeSession.isArchived.get();
 		disposables.add(autorun(reader => {
-			const isArchived = activeSession.isArchived.read(reader);
-			this._isActiveSessionArchived.set(isArchived);
-			if (isArchived && !wasArchived) {
-				this.openNewSessionView();
-			}
-			wasArchived = isArchived;
+			this._isActiveSessionArchived.set(activeSession.isArchived.read(reader));
 		}));
 
 		// Track workspace changes so the virtual-workspace context key stays in sync
 		disposables.add(autorun(reader => {
 			const workspace = activeSession.workspace.read(reader);
 			this._activeSessionWorkspaceIsVirtual.set(workspace?.isVirtualWorkspace ?? true);
-		}));
-
-		// Track chat list changes — if the active chat is removed, fall back
-		if (activeSession.status.get() !== SessionStatus.Untitled) {
-			disposables.add(autorun(reader => {
-				const chats = activeSession.chats.read(reader);
-				const activeChat = activeSession.activeChat.read(reader);
-				if (activeChat && !chats.some(c => this.uriIdentityService.extUri.isEqual(c.resource, activeChat.resource))) {
-					const fallback = chats[chats.length - 1] ?? activeSession.mainChat.read(reader);
-					if (fallback) {
-						this.openChat(activeSession, fallback.resource);
-					}
-				}
-			}));
-		}
-
-		// Track active chat changes to persist per-session state. The visible /
-		// active / sticky flags are snapshotted from the live grid at save time
-		// (see `_snapshotVisibleSessionStates`); here we only remember the last active
-		// chat so reopening the session restores its selected chat.
-		disposables.add(autorun(reader => {
-			const chat = activeSession.activeChat.read(reader);
-			if (chat && chat.status.read(undefined) !== SessionStatus.Untitled) {
-				const existing = this._sessionStates.get(activeSession.resource);
-				this._sessionStates.set(activeSession.resource, {
-					...existing,
-					sessionResource: activeSession.resource.toString(),
-					activeChatResource: chat.resource.toString(),
-				});
-			}
 		}));
 
 		return disposables;
@@ -327,8 +196,9 @@ export class SessionsManagementService extends Disposable implements ISessionsMa
 	}
 
 	private _handleDidReplaceSession(from: ISession, to: ISession): void {
-		this._visibility.updateSession(from, to);
 		this.chatWidgetHistoryService.moveHistory(ChatAgentLocation.Chat, from.sessionId, to.sessionId);
+		// Notify the view service so it can update the visible grid slot.
+		this._onDidReplaceSession.fire({ from, to });
 		// Always fire the change event so the SessionsList refreshes even when
 		// the user navigated to a different session while the new one was
 		// being created (which is how duplicate rows appeared in the list).
@@ -340,37 +210,18 @@ export class SessionsManagementService extends Disposable implements ISessionsMa
 	}
 
 	private onDidChangeSessionsFromSessionsProviders(e: ISessionChangeEvent): void {
-		this._onDidChangeSessions.fire(e);
-		const currentActive = this._visibility.activeSession.get();
-
-		// Clear stale pending session if the provider removed it. The provider
+		// Clear stale new session if the provider removed it. The provider
 		// already disposed it, so just drop the pointer (do not dispose again).
-		if (e.removed.length && this._pendingNewSession) {
-			if (e.removed.some(r => r.sessionId === this._pendingNewSession!.sessionId)) {
-				this._pendingNewSession = undefined;
-			}
-		}
-
-		// Clean removed sessions out of the visibility model (drops their grid slot
-		// and disposes their wrapper). If the active session is among the removed,
-		// removeMany picks a fallback active session (or clears it when no slot
-		// remains); drive the open flow below so the fallback is fully opened.
 		if (e.removed.length) {
-			this._visibility.removeMany(e.removed.map(r => r.sessionId));
-		}
-
-		if (!currentActive) {
-			return;
-		}
-
-		if (e.removed.length && e.removed.some(r => r.sessionId === currentActive.sessionId)) {
-			const fallback = this._visibility.activeSession.get();
-			if (fallback && this.getSession(fallback.resource)) {
-				this.openSession(fallback.resource);
-			} else {
-				this.openNewSessionView();
+			const current = this._newSession.get();
+			if (current && e.removed.some(r => r.sessionId === current.sessionId)) {
+				this._newSession.set(undefined, undefined);
 			}
 		}
+
+		// The view service reacts to this event to drop removed sessions from
+		// the grid and pick a fallback active session.
+		this._onDidChangeSessions.fire(e);
 	}
 
 	getSessions(): ISession[] {
@@ -385,37 +236,6 @@ export class SessionsManagementService extends Disposable implements ISessionsMa
 		return this.getSessions().find(s =>
 			this.uriIdentityService.extUri.isEqual(s.resource, resource)
 		);
-	}
-
-	getRecentlyOpenedSessions(): IRecentlyOpenedSessions {
-		const seen = new Set<string>();
-		const recent: ISession[] = [];
-
-		// Sessions in recency order (most-recently-opened first), deduplicated by
-		// session so a session with multiple opened chats appears only once and
-		// capped at the most recent {@link MAX_RECENTLY_OPENED_SESSIONS}.
-		for (const entry of this._recencyHistory.entries) {
-			if (recent.length >= MAX_RECENTLY_OPENED_SESSIONS) {
-				break;
-			}
-			const key = entry.sessionResource.toString();
-			if (seen.has(key)) {
-				continue;
-			}
-			seen.add(key);
-			const session = this.getSession(entry.sessionResource);
-			if (session) {
-				recent.push(session);
-			}
-		}
-
-		// Sessions that have not been included in the recently opened group,
-		// sorted by most recently updated first.
-		const other = this.getSessions()
-			.filter(s => !seen.has(s.resource.toString()))
-			.sort((a, b) => b.updatedAt.get().getTime() - a.updatedAt.get().getTime());
-
-		return { recent, other };
 	}
 
 	getAllSessionTypes(): ISessionType[] {
@@ -480,107 +300,33 @@ export class SessionsManagementService extends Disposable implements ISessionsMa
 		this._onDidChangeSessionTypes.fire();
 	}
 
-	/**
-	 * Cancel any in-flight open-session/restore and return a fresh cancellation token.
-	 */
-	private _startOpenSession() {
-		this._openSessionCts.value?.cancel();
-		const cts = new CancellationTokenSource();
-		this._openSessionCts.value = cts;
-		return cts.token;
-	}
-
-	/**
-	 * Cancel an in-flight {@link restoreVisibleSessions}. Called when the user
-	 * explicitly navigates to a specific session, so restore stops fighting
-	 * the user's choice. Additive new-session operations do NOT call this.
-	 */
-	private _cancelRestore(): void {
-		// `cancel()` (not just `clear()`/dispose) so the in-flight restore's
-		// token actually fires cancellation and bails out; `MutableDisposable`
-		// disposes the source without cancelling it.
-		this._restoreCts.value?.cancel();
-		this._restoreCts.clear();
-	}
-
-	async openChat(session: ISession, chatUri: URI): Promise<void> {
-		const t0 = Date.now();
-		this._cancelRestore();
-		const token = this._startOpenSession();
-		this.logService.trace(`[SessionsManagement] openChat start uri=${chatUri.toString()} provider=${session.providerId}`);
-		this.setActiveSession(session);
-		if (!await this._waitForSessionToLoad(session, token)) {
-			this.logService.trace(`[SessionsManagement] openChat cancelled while waiting for session to load uri=${chatUri.toString()}`);
+	setActiveSession(session: IActiveSession | undefined): void {
+		const previousSession = this._activeSession.get();
+		if (previousSession?.sessionId === session?.sessionId) {
 			return;
 		}
 
-		// Find the chat and update active chat
-		let chat: IChat | undefined;
-		//let previousChatResource: URI | undefined;
-		const activeSession = this._visibility.activeSession.get();
-		if (activeSession) {
-			//previousChatResource = activeSession.activeChat.get()?.resource;
-			chat = activeSession.chats.get().find(c => this.uriIdentityService.extUri.isEqual(c.resource, chatUri));
-			if (chat) {
-				this._visibility.setActiveChat(session, chat);
-			}
+		if (session) {
+			this.logService.info(`[SessionsManagement] Active session changed: ${session.resource.toString()}`);
+		} else {
+			this.logService.trace('[SessionsManagement] Active session cleared');
 		}
 
-		// If the chat is untitled (not yet sent), show the new-chat-in-session view
-		if (chat && chat.status.get() === SessionStatus.Untitled) {
-			this.logService.trace(`[SessionsManagement] openChat done total=${Date.now() - t0}ms uri=${chatUri.toString()} path=untitled`);
+		this._activeSession.set(session, undefined);
+	}
+
+	discardNewSession(session?: ISession): void {
+		const current = this._newSession.get();
+		if (!current) {
 			return;
 		}
-
-		this.logService.trace(`[SessionsManagement] openChat done total=${Date.now() - t0}ms uri=${chatUri.toString()}`);
-	}
-
-	async openSession(sessionResource: URI, options?: { preserveFocus?: boolean }): Promise<void> {
-		this._cancelRestore();
-		const token = this._startOpenSession();
-		await this._doOpenSession(sessionResource, token, options);
-	}
-
-	private async _doOpenSession(sessionResource: URI, token: CancellationToken, options?: { preserveFocus?: boolean }): Promise<void> {
-		const t0 = Date.now();
-		const sessionData = this.getSession(sessionResource);
-		if (!sessionData) {
-			this.logService.warn(`[SessionsManagement] openSession: session not found uri=${sessionResource.toString()}`);
-			throw new Error(`Session with resource ${sessionResource.toString()} not found`);
-		}
-		this.logService.trace(`[SessionsManagement] openSession start uri=${sessionResource.toString()} provider=${sessionData.providerId}`);
-		this.setActiveSession(sessionData);
-		if (!await this._waitForSessionToLoad(sessionData, token)) {
-			this.logService.trace(`[SessionsManagement] openSession cancelled while waiting for session to load uri=${sessionResource.toString()}`);
+		// When a specific session is given, only discard if it is the current
+		// new session; closing an unrelated session must not drop the draft.
+		if (session && session.sessionId !== current.sessionId) {
 			return;
 		}
-
-		this.logService.trace(`[SessionsManagement] openSession done total=${Date.now() - t0}ms uri=${sessionResource.toString()}`);
-	}
-
-	/**
-	 * Delete the currently pending (composed-but-not-sent) new session.
-	 *
-	 * Providers track new sessions themselves and no longer dispose them
-	 * implicitly when a newer one is created, so abandoning a pending session
-	 * must explicitly dispose it through its own provider to release the
-	 * eagerly-acquired backend session.
-	 *
-	 * This is only for abandonment. When a pending session is graduating (being
-	 * sent) or was already removed by its provider, just clear
-	 * {@link _pendingNewSession} directly so the session is left intact.
-	 */
-	private _deletePendingNewSession(): void {
-		const pending = this._pendingNewSession;
-		this._pendingNewSession = undefined;
-		if (pending) {
-			this._getProvider(pending)?.deleteNewSession(pending.sessionId);
-		}
-	}
-
-	unsetNewSession(): void {
-		this._deletePendingNewSession();
-		this.setActiveSession(undefined);
+		this._newSession.set(undefined, undefined);
+		this._getProvider(current)?.deleteNewSession(current.sessionId);
 	}
 
 	/**
@@ -629,30 +375,39 @@ export class SessionsManagementService extends Disposable implements ISessionsMa
 	}
 
 	createNewSession(folderUri: URI, options?: ICreateNewSessionOptions): ISession {
-		this._startOpenSession();
-
 		const { provider, sessionTypeId } = this._resolveProviderForNewSession(folderUri, options);
 
-		const previousPending = this._pendingNewSession;
+		const previousNewSession = this._newSession.get();
 		const session = provider.createNewSession(folderUri, sessionTypeId);
-		this._pendingNewSession = session;
-		this.setActiveSession(session);
+		this._newSession.set(session, undefined);
 
 		// Providers no longer dispose the previous new session implicitly, so
 		// dispose the one this composer just replaced. Use its own provider
 		// because switching workspace can switch providers. Done after a
 		// successful create so a throw above leaves the previous one intact.
-		if (previousPending && previousPending.sessionId !== session.sessionId) {
-			this._getProvider(previousPending)?.deleteNewSession(previousPending.sessionId);
+		if (previousNewSession && previousNewSession.sessionId !== session.sessionId) {
+			this._getProvider(previousNewSession)?.deleteNewSession(previousNewSession.sessionId);
 		}
 		return session;
+	}
+
+	async createNewChatInSession(session: ISession): Promise<IChat | undefined> {
+		const provider = this._getProvider(session);
+		if (!provider) {
+			this.logService.warn(`[SessionsManagement] createNewChatInSession: provider '${session.providerId}' not found`);
+			return undefined;
+		}
+		// Reuse an existing untitled chat if one exists, otherwise create a new one.
+		const existingUntitled = session.chats.get().find(c => c.status.get() === SessionStatus.Untitled);
+		return existingUntitled ?? await provider.createNewChat(session.sessionId);
 	}
 
 	async sendNewChatRequest(session: ISession, options: ISendRequestOptions): Promise<void> {
 		// The session is graduating into the list (being sent),
 		// so the provider keeps owning it — just drop the pointer, do not delete.
-		this._pendingNewSession = undefined;
-		this._isNewChatSessionContext.set(false);
+		// Clearing the new session recomputes the isNewChatSession context key
+		// via the active-session autorun.
+		this._newSession.set(undefined, undefined);
 
 		const provider = this._getProvider(session);
 		if (!provider) {
@@ -660,14 +415,11 @@ export class SessionsManagementService extends Disposable implements ISessionsMa
 		}
 
 		if (options.background) {
-			// Restore the new-session view synchronously so the composer stays
-			// put, then run the send fire-and-forget so the composer can reset
-			// and reseed immediately while the request commits. If the commit
-			// fails, the graduating draft is stranded (no longer in
-			// `_pendingNewSession`), so dispose it through its provider to
-			// release the eager backend session. Safe no-op if the provider
-			// already graduated/removed it.
-			this.openNewSessionView();
+			// Run the send fire-and-forget so the composer can reset and reseed
+			// immediately while the request commits. If the commit fails, the
+			// graduating draft is stranded (no longer the current new session),
+			// so dispose it through its provider to release the eager backend
+			// session. Safe no-op if the provider already graduated/removed it.
 			this._sendNewChatRequestInBackground(provider, session, options).catch(e => {
 				provider.deleteNewSession(session.sessionId);
 				this.logService.error('[SessionsManagement] Failed to send background request:', e);
@@ -678,53 +430,27 @@ export class SessionsManagementService extends Disposable implements ISessionsMa
 		// Foreground send: notify listeners that a send is starting. Listeners
 		// (e.g., telemetry) can use this to prewarm caches whose result is
 		// consumed when `onDidSendRequest` fires below. The background path
-		// fires this from within `_sendNewChatRequestInBackground`.
+		// fires this from within `_sendNewChatRequestInBackground`. The view
+		// service observes the will/did send pair to keep the newest chat
+		// active in the visible slot while the send materialises.
 		this._onWillSendRequest.fire(session);
 
-		const setActiveChatToLast = () => {
-			const activeSession = this._visibility.activeSession.get();
-			if (activeSession?.sessionId === session.sessionId && this.uriIdentityService.extUri.isEqual(activeSession.activeChat.get().resource, (<IActiveSession>session).activeChat?.get().resource)) {
-				const chats = activeSession.chats.get();
-				const lastChat = chats[chats.length - 1];
-				if (lastChat) {
-					this._visibility.setActiveChat(session, lastChat);
-				}
-			}
-		};
+		// Ask the provider to create the new chat, then send the request.
+		const chat = await provider.createNewChat(session.sessionId, options.query);
 
-		// Listen for chats changing during the send (subsequent chat appears in the group)
-		const chatsListener = autorun(reader => {
-			session.chats.read(reader);
-			setActiveChatToLast();
-		});
-
+		const chatResourceKey = chat.resource.toString();
+		this._pendingSendChatResources.add(chatResourceKey);
+		let updatedSession: ISession;
 		try {
-			// Ask the provider to create the new chat; open its widget before sending
-			const chat = await provider.createNewChat(session.sessionId, options.query);
-			// Swap in a transient session whose resource is the new chat
-			// resource, so the grid slot reflects the chat that is about to
-			// be sent before the provider hands us the final session.
-			const tmpSession = this._visibility.updateResourceOfSession(session, chat.resource);
-
-			const chatResourceKey = chat.resource.toString();
-			this._pendingSendChatResources.add(chatResourceKey);
-			let updatedSession: ISession;
-			try {
-				updatedSession = await provider.sendRequest(session.sessionId, chat.resource, options);
-			} finally {
-				this._pendingSendChatResources.delete(chatResourceKey);
-			}
-			if (updatedSession.sessionId !== session.sessionId) {
-				this.logService.info(`[SessionsManagement] sendRequest: active session replaced: ${session.sessionId} -> ${updatedSession.sessionId}`);
-				this._visibility.updateSession(tmpSession, updatedSession);
-				setActiveChatToLast();
-			}
-			this._onDidStartSession.fire(updatedSession);
-
-			this._onDidSendRequest.fire({ session: updatedSession, chat, isNewSession: true, isNewChat: true, options });
+			updatedSession = await provider.sendRequest(session.sessionId, chat.resource, options);
 		} finally {
-			chatsListener.dispose();
+			this._pendingSendChatResources.delete(chatResourceKey);
 		}
+		if (updatedSession.sessionId !== session.sessionId) {
+			this.logService.info(`[SessionsManagement] sendRequest: active session replaced: ${session.sessionId} -> ${updatedSession.sessionId}`);
+		}
+		this._onDidStartSession.fire(updatedSession);
+		this._onDidSendRequest.fire({ session: updatedSession, chat, isNewSession: true, isNewChat: true, options });
 	}
 
 	/**
@@ -797,17 +523,15 @@ export class SessionsManagementService extends Disposable implements ISessionsMa
 	}
 
 	async sendRequest(session: ISession, chat: IChat, options: ISendRequestOptions): Promise<void> {
-		// Sending into an existing session abandons any pending composer, so
-		// dispose it to release its eager backend session.
-		this._deletePendingNewSession();
+		// Sending into an existing session abandons any in-progress new session,
+		// so dispose it to release its eager backend session.
+		this.discardNewSession();
 
 		// Notify listeners that a send is starting. Listeners (e.g., telemetry)
 		// can use this to prewarm caches whose result is consumed when
-		// `onDidSendRequest` fires below.
+		// `onDidSendRequest` fires below. The view service observes the will/did
+		// send pair to keep the sent chat active in the visible slot.
 		this._onWillSendRequest.fire(session);
-
-		// Keep the sent chat as the active chat
-		this._visibility.setActiveChat(session, chat);
 
 		const provider = this._getProvider(session);
 		if (!provider) {
@@ -824,464 +548,9 @@ export class SessionsManagementService extends Disposable implements ISessionsMa
 		}
 		if (updatedSession.sessionId !== session.sessionId) {
 			this.logService.info(`[SessionsManagement] sendRequest: active session replaced: ${session.sessionId} -> ${updatedSession.sessionId}`);
-			this._visibility.updateSession(session, updatedSession);
 		}
 
 		this._onDidSendRequest.fire({ session: updatedSession, chat, isNewSession: false, isNewChat: true, options });
-	}
-
-	openNewSessionView(options?: { inheritWorkspaceFromActiveSession?: boolean }): void {
-		const current = this._visibility.activeSession.get();
-		// No-op when no session is active (empty new-session placeholder showing).
-		if (current === undefined) {
-			return;
-		}
-		this._startOpenSession();
-
-		// When explicitly requested, inherit the workspace of the session being
-		// switched away from so the new session view opens in the same workspace
-		// rather than defaulting to the last composed new session's workspace.
-		// Only inherit from an established session (not the pending new session)
-		// and only when its workspace differs from any pending new session, so
-		// an existing in-progress draft for that same workspace is preserved.
-		if (options?.inheritWorkspaceFromActiveSession && current.sessionId !== this._pendingNewSession?.sessionId) {
-			const inheritUri = current.workspace.get()?.folders[0]?.root;
-			const pendingUri = this._pendingNewSession?.workspace.get()?.folders[0]?.root;
-			if (inheritUri && !this.uriIdentityService.extUri.isEqual(inheritUri, pendingUri)) {
-				try {
-					// Creates a fresh pending new session for the inherited
-					// workspace and makes it active, disposing the previous one.
-					this.createNewSession(inheritUri);
-					this._onDidOpenNewSessionView.fire();
-					return;
-				} catch (e) {
-					this.logService.warn(`[SessionsManagement] openNewSessionView: failed to inherit workspace '${inheritUri.toString()}', falling back to default`, e);
-				}
-			}
-		}
-
-		// Restore the pending new session if one exists, so pickers
-		// re-derive their state from the still-alive session object.
-		// Otherwise clear active session (first time / after send).
-		this.setActiveSession(this._pendingNewSession ?? undefined);
-		this._onDidOpenNewSessionView.fire();
-	}
-
-	async openNewChatInSession(session: ISession): Promise<void> {
-		this._cancelRestore();
-		this._startOpenSession();
-		const provider = this._getProvider(session);
-		if (!provider) {
-			this.logService.warn(`[SessionsManagement] openNewChatInSession: provider '${session.providerId}' not found`);
-			return;
-		}
-
-		// Reuse an existing untitled chat if one exists, otherwise create a new one
-		const existingUntitled = session.chats.get().find(c => c.status.get() === SessionStatus.Untitled);
-		const chat = existingUntitled ?? await provider.createNewChat(session.sessionId);
-
-		this.setActiveSession(session);
-
-		// Set the chat as the active chat
-		this._visibility.setActiveChat(session, chat);
-	}
-
-	setActive(session: IActiveSession | undefined) {
-		this.setActiveSession(session);
-	}
-
-	private setActiveSession(session: ISession | undefined, force?: boolean): void {
-		const previousSession = this._visibility.activeSession.get();
-		if (!force && previousSession?.sessionId === session?.sessionId) {
-			return;
-		}
-
-		if (session) {
-			this.logService.info(`[ActiveSessionService] Active session changed: ${session.resource.toString()}`);
-		} else {
-			this.logService.trace('[ActiveSessionService] Active session cleared');
-		}
-
-		this._visibility.setActive(session);
-	}
-
-	toggleSessionStickiness(session: ISession): void {
-		const sticky = this._visibility.toggleStickiness(session);
-		this._onDidToggleSessionStickiness.fire({ session, sticky });
-	}
-
-	insertAt(session: ISession, targetSessionId: string, side: 'left' | 'right', activate: boolean = true): void {
-		this._visibility.insertAt(session, targetSessionId, side, activate);
-	}
-
-	closeSession(session: ISession | undefined): void {
-		const sessionId = session?.sessionId;
-		const visible = this._visibility.visibleSessions.get();
-		if (!visible.some(s => s?.sessionId === sessionId)) {
-			return;
-		}
-
-		// The empty/new-session slot has no sessionId; both it and "no active
-		// session" are reported by activeSession as undefined. Since we already
-		// confirmed the slot is present in `visible`, undefined === undefined
-		// here means the empty slot is active.
-		const activeSessionId = this._visibility.activeSession.get()?.sessionId;
-		const wasActive = activeSessionId === sessionId;
-
-		if (sessionId === undefined || this._pendingNewSession?.sessionId === sessionId) {
-			this._deletePendingNewSession();
-		}
-
-		this._visibility.removeMany([sessionId]);
-
-		if (!wasActive) {
-			return;
-		}
-
-		// removeMany already picked a fallback active session (or cleared the
-		// active observable when no slot remains); drive the full open flow.
-		const fallback = this._visibility.activeSession.get();
-		if (fallback === undefined) {
-			this.openNewSessionView();
-		}
-	}
-
-	closeAllSessions(): void {
-		const ids = this._visibility.visibleSessions.get()
-			.filter((s): s is IActiveSession => !!s)
-			.map(s => s.sessionId);
-		if (ids.length === 0) {
-			return;
-		}
-
-		this._pendingNewSession = undefined;
-
-		// Remove every visible session in a single pass; the visibility model
-		// clears the active session, which drives the grid back to the
-		// new-session view via the part's reconciliation.
-		this._visibility.removeMany(ids);
-	}
-
-	private _restoreInitialChat(session: ISession): IChat {
-		const chats = session.chats.get();
-		let initialChat = chats[0];
-		const sessionState = this._sessionStates.get(session.resource);
-		if (sessionState?.activeChatResource) {
-			try {
-				const lastChatResource = URI.parse(sessionState.activeChatResource);
-				const found = chats.find(c => this.uriIdentityService.extUri.isEqual(c.resource, lastChatResource));
-				if (found) {
-					initialChat = found;
-				}
-			} catch (error) {
-				this.logService.warn('[ActiveSessionService] Failed to restore active chat from stored session state', error);
-			}
-		}
-		return initialChat;
-	}
-
-	private async _waitForSessionToLoad(session: ISession, token: CancellationToken): Promise<boolean> {
-		if (!session.loading.get()) {
-			return true;
-		}
-		if (token.isCancellationRequested) {
-			return false;
-		}
-
-		await new Promise<void>(resolve => {
-			const disposables = new DisposableStore();
-			let resolved = false;
-			const finish = () => {
-				if (resolved) {
-					return;
-				}
-				resolved = true;
-				disposables.dispose();
-				resolve();
-			};
-
-			disposables.add(token.onCancellationRequested(finish));
-			disposables.add(autorun(reader => {
-				if (!session.loading.read(reader)) {
-					finish();
-				}
-			}));
-		});
-
-		return !token.isCancellationRequested;
-	}
-
-	private _loadSessionStates(): ResourceMap<ISessionState> {
-		const map = new ResourceMap<ISessionState>();
-		const raw = this.storageService.get(ACTIVE_SESSION_STATES_KEY, StorageScope.WORKSPACE);
-		if (!raw) {
-			return map;
-		}
-		try {
-			const entries: ISessionState[] = JSON.parse(raw);
-			for (const entry of entries) {
-				const uri = URI.parse(entry.sessionResource);
-				map.set(uri, entry);
-			}
-		} catch {
-			// ignore corrupt data
-		}
-		return map;
-	}
-
-	private _saveSessionStates(): void {
-		const entries = this._snapshotVisibleSessionStates();
-		this.storageService.store(ACTIVE_SESSION_STATES_KEY, JSON.stringify(entries), StorageScope.WORKSPACE, StorageTarget.MACHINE);
-	}
-
-	private _snapshotVisibleSessionStates(): ISessionState[] {
-		const activeId = this._visibility.activeSession.get()?.sessionId;
-		const visible = this._visibility.visibleSessions.get();
-		const entries: ISessionState[] = [];
-		visible.forEach((session, index) => {
-			if (!session) {
-				return;
-			}
-
-			if (session.status.get() === SessionStatus.Untitled) {
-				this._sessionStates.delete(session.resource);
-				return;
-			}
-
-			// Keep the in-memory record up to date so the session's last active
-			// chat is remembered while reopening it within this window.
-			const existing = this._sessionStates.get(session.resource);
-			const state: ISessionState = {
-				sessionResource: session.resource.toString(),
-				activeChatResource: session.activeChat.get()?.resource.toString() ?? existing?.activeChatResource,
-				visibleOrder: index,
-				isSticky: session.sticky.get(),
-				isActive: session.sessionId === activeId,
-			};
-			this._sessionStates.set(session.resource, state);
-			entries.push(state);
-		});
-		return entries;
-	}
-
-	/**
-	 * The persisted visible sessions, ordered left-to-right by their stored
-	 * grid position.
-	 */
-	private _getVisibleSessionStates(): ISessionState[] {
-		const states: ISessionState[] = [];
-		for (const [, state] of this._sessionStates) {
-			if (state.visibleOrder !== undefined) {
-				states.push(state);
-			}
-		}
-		return states.sort((a, b) => (a.visibleOrder! - b.visibleOrder!));
-	}
-
-	/**
-	 * Wait for the session with the given resource to become available via its
-	 * provider, resolving with the session or `undefined` if the token is
-	 * cancelled before it appears. When `timeout` is given, resolves with
-	 * `undefined` after that many milliseconds so a persisted session that never
-	 * resurfaces (e.g. deleted while the window was closed) cannot keep restore
-	 * pending — and its provider listeners alive — indefinitely.
-	 */
-	private _waitForSession(sessionResource: URI, token: CancellationToken, timeout?: number): Promise<ISession | undefined> {
-		const existing = this.getSession(sessionResource);
-		if (existing) {
-			return Promise.resolve(existing);
-		}
-		return new Promise<ISession | undefined>(resolve => {
-			const disposables = new DisposableStore();
-			let resolved = false;
-			const finish = (session: ISession | undefined) => {
-				if (resolved) {
-					return;
-				}
-				resolved = true;
-				disposables.dispose();
-				resolve(session);
-			};
-
-			disposables.add(token.onCancellationRequested(() => finish(undefined)));
-
-			const tryFind = () => {
-				if (token.isCancellationRequested) {
-					finish(undefined);
-					return;
-				}
-				const session = this.getSession(sessionResource);
-				if (session) {
-					finish(session);
-				}
-			};
-
-			// Providers (e.g. the agent host) load their session cache
-			// asynchronously, so the session may appear via either a provider
-			// change or a session list change.
-			disposables.add(this.sessionsProvidersService.onDidChangeProviders(() => tryFind()));
-			disposables.add(this.onDidChangeSessions(() => tryFind()));
-
-			// Give up after the timeout so the listeners above are not retained
-			// forever when the session is gone for good.
-			if (timeout !== undefined) {
-				disposables.add(disposableTimeout(() => finish(undefined), timeout));
-			}
-
-			// In case the session became available between the initial check and
-			// the listener registration.
-			tryFind();
-		});
-	}
-
-	async restoreVisibleSessions(): Promise<void> {
-		// Ordered list of slots to restore: real sessions plus, optionally, the
-		// empty (new-session) slot when it was active.
-		interface IRestoreTarget {
-			readonly resource: URI | undefined;
-			readonly isSticky: boolean;
-			readonly isActive: boolean;
-			readonly order: number;
-		}
-
-		const targets: IRestoreTarget[] = this._getVisibleSessionStates().map(state => ({
-			resource: URI.parse(state.sessionResource),
-			isSticky: !!state.isSticky,
-			isActive: !!state.isActive,
-			order: state.visibleOrder!,
-		}));
-
-		if (targets.length === 0) {
-			targets.push({ resource: undefined, isSticky: false, isActive: true, order: 1 });
-		}
-
-		targets.sort((a, b) => a.order - b.order);
-
-		let activeIdx = targets.findIndex(t => t.isActive);
-		if (activeIdx < 0) {
-			activeIdx = 0;
-		}
-
-		// Use a dedicated cancellation token (not the shared open-session one)
-		// so that a new-session draft created during restore (e.g. by the
-		// new-chat composer on startup) does not abort restoring the grid. The
-		// token is cancelled only when the user explicitly opens a session.
-		const cts = new CancellationTokenSource();
-		this._restoreCts.value = cts;
-		const token = cts.token;
-
-		// Sessions resolved so far, indexed by their position in `targets`.
-		// `null` marks the empty (new-session) slot, which has no session.
-		const resolved: (ISession | null | undefined)[] = new Array(targets.length).fill(undefined);
-
-		/**
-		 * Insert a resolved session into the grid next to the nearest
-		 * already-placed neighbour, preserving the persisted order regardless of
-		 * the order in which sessions become available. When a neighbour exists
-		 * the active session is left unchanged; only in the edge case where no
-		 * neighbour has been placed yet (e.g. the active target never resurfaced,
-		 * so the grid laid out empty) does the first session to arrive become
-		 * active as a sensible fallback.
-		 */
-		const place = (idx: number, session: ISession): void => {
-			let anchor: { id: string | undefined; side: 'left' | 'right' } | undefined;
-			for (let j = idx - 1; j >= 0 && !anchor; j--) {
-				const neighbour = resolved[j];
-				if (neighbour !== undefined) {
-					anchor = { id: neighbour?.sessionId, side: 'right' };
-				}
-			}
-			for (let j = idx + 1; j < targets.length && !anchor; j++) {
-				const neighbour = resolved[j];
-				if (neighbour !== undefined) {
-					anchor = { id: neighbour?.sessionId, side: 'left' };
-				}
-			}
-
-			resolved[idx] = session;
-			if (anchor) {
-				this._visibility.insertAt(session, anchor.id, anchor.side, false);
-			} else {
-				this.setActiveSession(session);
-			}
-			if (targets[idx].isSticky) {
-				this._visibility.toggleStickiness(session);
-			}
-		};
-
-		// Resolve the active session first so it can act as the anchor for the
-		// initial layout. The empty slot resolves immediately (the grid already
-		// shows the new-session view). Load progress is surfaced per-leaf by the
-		// chat view itself once the grid is laid out (mirroring how each editor
-		// group owns its progress bar), so no part-wide progress is driven here.
-		const activeTarget = targets[activeIdx];
-		const activeSessionPromise: Promise<ISession | undefined> = activeTarget.resource
-			? this._waitForSession(activeTarget.resource, token, RESTORE_SESSION_WAIT_TIMEOUT).then(session => session ?? undefined)
-			: Promise.resolve<ISession | undefined>(undefined);
-
-		const activeSession = await activeSessionPromise;
-
-		if (token.isCancellationRequested) {
-			return;
-		}
-
-		// Lay out all currently-available sessions atomically in the persisted
-		// order so the grid appears in one shot rather than building up slot by
-		// slot (which caused the active session to be shown alone and then
-		// reflow as the others were inserted). Sessions whose provider has not
-		// yet surfaced them are filled in incrementally below.
-		const slots: { session: ISession | undefined; sticky: boolean }[] = [];
-		let activeSlotIndex = -1;
-		for (let idx = 0; idx < targets.length; idx++) {
-			const target = targets[idx];
-			let session: ISession | null | undefined;
-			if (!target.resource) {
-				session = null; // empty new-session slot
-			} else if (idx === activeIdx) {
-				session = activeSession;
-			} else {
-				session = this.getSession(target.resource);
-			}
-			if (session === undefined) {
-				continue; // not yet available — placed incrementally below
-			}
-			resolved[idx] = session;
-			if (idx === activeIdx) {
-				activeSlotIndex = slots.length;
-			}
-			slots.push({ session: session ?? undefined, sticky: target.isSticky });
-		}
-		this._visibility.restoreGrid(slots, activeSlotIndex);
-
-		if (token.isCancellationRequested) {
-			return;
-		}
-
-		// Focus is moved into the restored active session by the sessions part,
-		// which observes the active-session change (see SessionsPartService).
-
-		// Place any sessions that became available later in their correct
-		// positions around the already-established layout.
-		await Promise.all(targets.map(async (target, idx) => {
-			if (idx === activeIdx || !target.resource || token.isCancellationRequested || resolved[idx] !== undefined) {
-				return;
-			}
-			const session = await this._waitForSession(target.resource, token, RESTORE_SESSION_WAIT_TIMEOUT);
-			if (!session || token.isCancellationRequested || resolved[idx] !== undefined) {
-				return;
-			}
-			place(idx, session);
-		}));
-	}
-
-	// -- Session Navigation --
-
-	async openPreviousSession(): Promise<void> {
-		await this._navigation.goBack();
-	}
-
-	async openNextSession(): Promise<void> {
-		await this._navigation.goForward();
 	}
 
 	// -- Session Actions --
