+++ src/vs/sessions/sessions.common.main.ts
@@ -447,7 +447,8 @@ import '../workbench/contrib/opener/browser/opener.contribution.js';
 
 import './browser/paneCompositePartService.js';
 import './browser/parts/editorParts.js';
-import './browser/parts/sessionsPartService.js';
+import './browser/parts/sessionsParts.js';
+import './browser/sessionsViewService.js';
 import './browser/parts/menubar.contribution.js';
 import './browser/layoutActions.js';
